basic ui setup, for instruction and image capture
