+++ package.json
@@ -3,18 +3,19 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@tensorflow-models/body-pix": "^2.0.5",
-    "@tensorflow/tfjs": "^3.0.0",
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
+    "@tensorflow-models/body-pix": "^2.1.0",
+    "@tensorflow/tfjs": "^3.5.0",
+    "@testing-library/jest-dom": "^5.12.0",
+    "@testing-library/react": "^11.2.6",
+    "@testing-library/user-event": "^12.8.3",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
     "react-fontawesome": "^1.7.1",
     "react-material-iconic-font": "^2.0.0",
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.1",
-    "react-webcam": "^5.2.2",
+    "react-step-wizard": "^5.3.6",
+    "react-webcam": "^5.2.3",
     "web-vitals": "^0.2.4",
     "workbox-background-sync": "^5.1.3",
     "workbox-broadcast-update": "^5.1.3",
